Add post method
Add post method that contains a conditional which states IF the value of the form  search button is
"Search", use the search_rewards_data_url endpoint to search the DB collection for the document matching the supplied
email address,  and set the context to rewards_data and all_rewards_data. Otherwise,
use the send_order_data_url endpoint to accept an email address and order total to send to the DB collection.