Display RewardsUI
In the RewardsUI, I was able to create a table that displays the reward data for all users. I was also able to make it so when a user enters the customer's order data or filter for a specific customer, the api endpoint previously made is called and the data is displayed but not in the html page.